+++ CHANGELOG.md
@@ -40,6 +40,7 @@ for further information about branching and tagging conventions.
 - Update asset requests page & module
 - Create sale requests module
 - Pre-issuance requests module
+- Incoming withdrawal requests module
 - LOAD_KV_ENTRIES action to key-values vuex module
 - Subject info `Dividend for <token-name>` for every payment operation in dividends
 - `All opportunities(All sales)` and `My opportunities(My sales)` pages on the `Opportunities(Funds)` page
@@ -84,6 +85,8 @@ for further information about branching and tagging conventions.
 - Change defaultQuoteAsset to single acceptable asset for the Revenue opportunities
 - "Tokens" page responsive design
 - Do not allow unverified users to add an asset with KYC required type to their balances
+- Composing creator details on withdrawal form (use address form master 
+account & comment for other ones)
 
 ## [1.3.1-rc.0] - 2019-03-20